Add direct broken-tool lookup for self-repair
Replace the post-claim repair-attempt reload with a by-name tool failure
lookup so self-repair no longer scans every broken tool before enforcing
attempt limits.

Move the default self-repair helper methods into a sibling module so
`default.rs` stays below the project file-length threshold without changing
repair behaviour.
